+++ CHANGELOG.md
@@ -2,10 +2,13 @@
 
 - **BREAKING** No longer output `strings.txt` with smalisa output. This has been reworked to include strings in the graph database so they can be queried alongside the graph. It is still possible to get all strings for a given source, but this file no longer exists
 - **BREAKING** Updated the GraphDatabase trait to include string and field related methods
+- **BREAKING** Update to apktool v3.0.2 -- this is breaking because the new apktool doesn't allow specifying `-api`
 - Major bugfixes:
     - Permissions were UNIQUE(name), updated to UNIQUE(name, source_apk_id)
     - Missing index on a graph query made it take very long
 - Surface method IDs in graph queries
+- Updated smali/baksmali/jadx versions in the nix flake
+- Add `DTU_DIFF_SOURCE` env var for specifying a default diff source
 - Added `find strings` to find strings by various criteria
 - Added `find methods` to find methods by various criteria
 - Added `find fields` to find fields by various criteria